In wfa_cmds.h, the framerate in tgPingStart_t struct is declared as float type instead of int to resolve the issue in SIG-523.
SCR#SIG-523
close[s:jira, t:523]

+++ inc/wfa_cmds.h
@@ -66,7 +66,12 @@ typedef struct _tg_ping_start
 {
     char dipaddr[IPV6_ADDRESS_STRING_LEN];  /* destination/remote ip address */
     int  frameSize;
-    int  frameRate;
+    //<<<----------------------------------------------------------------------------------
+    //jira issue: SIG-523
+    //define it as float type to resolve the rounding issue when the framerate is not int
+    float  frameRate;
+    ////int  frameRate;
+    //---------------------------------------------------------------------------------->>>
     int  duration;
     int  type;
     int  qos;
